feat: lower case components
