+++ app/components/dialogs/recibos/create-recibo.tsx
@@ -4,9 +4,9 @@ import { type SubmitHandler, useForm } from 'react-hook-form'
 import { z } from 'zod'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { type Farm } from '~/utils/types'
-import { Select } from '../../Form/Select'
-import { Input } from '../../Form/Input'
-import { TextArea } from '../../Form/TextArea'
+import { Select } from '../../form/select'
+import { Input } from '../../form/input'
+import { TextArea } from '../../form/text-area'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { X } from 'phosphor-react'
 import { createReceipt } from '~/utils/api/create-receipt'
