+++ app/components/dialogs/recibos/edit-recibo.tsx
@@ -3,8 +3,8 @@ import { useEffect, useState } from 'react'
 import { type SubmitHandler, useForm } from 'react-hook-form'
 import { z } from 'zod'
 import { zodResolver } from '@hookform/resolvers/zod'
-import { Input } from '../../Form/Input'
-import { TextArea } from '../../Form/TextArea'
+import { Input } from '../../form/input'
+import { TextArea } from '../../form/text-area'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { type Receipt } from '~/utils/types'
 import { Pencil, X } from 'phosphor-react'
