+++ app/routes/index.tsx
@@ -1,12 +1,12 @@
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
 import { useState } from 'react'
 import { Printer, Trash } from 'phosphor-react'
-import { Pagination } from '~/components/Pagination'
-import { Container } from '~/components/Container'
-import { CreateReciboDialog } from '~/components/Dialogs/Recibos/CreateRecibo'
-import { PrintListagem } from '~/components/Dialogs/Recibos/PrintListagem'
-import { PrintRecibos } from '~/components/Dialogs/Recibos/PrintRecibos'
-import { EditReciboDialog } from '~/components/Dialogs/Recibos/EditRecibo'
+import { Pagination } from '~/components/pagination'
+import { Container } from '~/components/container'
+import { CreateReciboDialog } from '~/components/dialogs/recibos/create-recibo'
+import { PrintListagem } from '~/components/dialogs/recibos/print-listagem'
+import { PrintRecibos } from '~/components/dialogs/recibos/print-recibos'
+import { EditReciboDialog } from '~/components/dialogs/recibos/edit-recibo'
 import { getReceipts } from '~/utils/api/get-receipts'
 import { getFarms } from '~/utils/api/get-farms'
 import { deleteReceipt } from '~/utils/api/delete-receipt'
